requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
apsw==3.48.0.0
apswutils==0.0.2
apt-xapian-index==0.49
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.16.0
beautifulsoup4==4.12.2
beniget==0.4.1
bleach==6.2.0
blessed==1.20.0
blinker==1.4
Brotli==1.0.9
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
csvkit==1.0.6
cupshelpers==1.0
cycler==0.12.1
dbfread==2.0.7
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
deepsearch-glm==0.26.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
docling==2.5.2
docling-core==2.3.2
docling-ibm-models==2.0.3
docling-parse==2.0.4
docutils==0.21.2
duckduckgo_search==5.3.1b1
easyocr==1.7.2
editor==1.6.6
emoji==2.12.1
et-xmlfile==1.0.1
eval_type_backport==0.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastcore==1.7.28
fastjsonschema==2.21.1
fastlite==0.1.1
filelock==3.13.1
filetype==1.2.0
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.12
fsspec==2023.12.2
galternatives==1.0.8
gast==0.5.2
GDAL==3.4.1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-genai==0.2.2
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gpt4all==2.7.0
gradio_client==1.3.0
greenlet==1.1.2
groq==0.11.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2image==2.0.4.3
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.6
ijson==3.3.0
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib-metadata==7.0.0
iniconfig==1.1.1
inquirer==3.1.4
ipykernel==6.27.1
ipython==8.18.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jdcal==1.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.10.0
jsonlines==3.1.0
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
leather==0.3.4
litellm==0.13.2
lxml==4.9.4
lz4==3.1.3+dfsg
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mean-average-precision==2021.4.26.0
meson==0.61.2
mistralai==1.1.0
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mutagen==1.45.1
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.8.1
notebook==7.3.1
notebook_shim==0.2.4
notify2==0.3.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
ollama==0.2.1
open-interpreter==0.1.17
openai==1.36.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.0.9
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.1.0
pluggy==0.13.0
ply==3.11
polars==1.2.1
primp==0.8.0
proglog==0.1.10
prometheus_client==0.21.1
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==5.29.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyaes==1.6.1
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycairo==1.20.1
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.17.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyOpenGL==3.1.5
pyparsing==2.4.7
pypdfium2==4.30.0
PyQt5-sip==12.9.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==6.2.5
python-apt==2.4.0+ubuntu3
python-bidi==0.6.3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.0
python-editor==1.0.4
python-fasthtml==0.12.0
python-json-logger==3.2.1
python-multipart==0.0.20
python-pptx==1.0.2
python-slugify==4.0.0
pythran==0.10.0
pytimeparse==1.1.5
pytube==15.0.0
pytz==2024.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
readchar==4.0.5
referencing==0.35.1
regex==2023.10.3
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.21.0
rsa==4.7.2
Rtree==1.3.0
runs==1.2.2
scikit-image==0.24.0
scipy==1.14.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.16.0
Selenium-Screenshot==2.1.0
Send2Trash==1.8.3
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.31
stack-data==0.6.3
starlette==0.45.2
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
Telethon==1.21.1
termcolor==2.4.0
terminado==0.18.1
tifffile==2024.9.20
tiktoken==0.4.0
tinycss2==1.4.0
together==1.3.0
tokenizers==0.15.0
tokentrim==0.1.13
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.0
trio==0.23.2
trio-websocket==0.11.1
triton==3.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
Unidecode==1.3.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
usb-creator==0.3.7
uvicorn==0.34.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.4
wcwidth==0.2.12
webcolors==24.11.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==14.1
wget==3.2
wsproto==1.2.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.2.0
xmod==1.8.1
yarl==1.12.1
yaspin==3.0.1
youtube-dl==2021.12.17
yt-dlp==2022.4.8
zipp==1.0.0
